+++ tests/testthat/test-show_dplyr.R
@@ -54,9 +54,7 @@ test_that("show_dplyr() on SELECT example query #3 returns expected result", {
 
 test_that("show_dplyr() truncates function arguments after the fifth argument with older versions of rlang", {
   skip_if_not(exists("games"), message = "Test data not loaded")
-  # uncomment this line after a version of rlang with
-  # https://github.com/r-lib/rlang/pull/897 merged is on CRAN
-  # skip_if("max_elements" %in% names(formals(rlang::expr_deparse)))
+  skip_if("max_elements" %in% names(formals(rlang::expr_deparse)))
   expect_equal(
     str2lang(paste(trimws(capture.output(
       show_dplyr("SELECT id, name, inventor, year, min_age, min_players, max_players, list_price FROM games")
@@ -65,15 +63,13 @@ test_that("show_dplyr() truncates function arguments after the fifth argument wi
   )
 })
 
-# uncomment this test after a version of rlang with
-# https://github.com/r-lib/rlang/pull/897 merged is on CRAN
-# test_that("show_dplyr() does not truncate function arguments with newer versions of rlang", {
-#   skip_if_not(exists("games"), message = "Test data not loaded")
-#   skip_if_not("max_elements" %in% names(formals(rlang::expr_deparse)))
-#   expect_equal(
-#     str2lang(paste(trimws(capture.output(
-#       show_dplyr("SELECT id, name, inventor, year, min_age, min_players, max_players, list_price FROM games")
-#     )), collapse = " ")),
-#     quote(games %>% select(id, name, inventor, year, min_age, min_players, max_players, list_price))
-#   )
-# })
+test_that("show_dplyr() does not truncate function arguments with newer versions of rlang", {
+  skip_if_not(exists("games"), message = "Test data not loaded")
+  skip_if_not("max_elements" %in% names(formals(rlang::expr_deparse)))
+  expect_equal(
+    str2lang(paste(trimws(capture.output(
+      show_dplyr("SELECT id, name, inventor, year, min_age, min_players, max_players, list_price FROM games")
+    )), collapse = " ")),
+    quote(games %>% select(id, name, inventor, year, min_age, min_players, max_players, list_price))
+  )
+})